# internal/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_projecthelena_warden_internal_db.CheckResult:
    properties:
      latency:
        type: integer
      monitorId:
        type: string
      status:
        type: string
      statusCode:
        type: integer
      timestamp:
        type: string
    type: object
...
  internal_api.GroupDTO:
    properties:
      id:
        type: string
      monitors:
        items:
          $ref: '#/definitions/internal_api.MonitorDTO'
        type: array
      name:
        type: string
    type: object
  internal_api.GroupOverviewDTO:
    properties:
      id:
        type: string
      name:
        type: string
      status:
        description: up, down, degraded
        type: string
    type: object
  internal_api.HistoryPoint:
    properties:
      latency:
        type: integer
      status:
        type: string
      statusCode:
        type: integer
      timestamp:
        type: string
    type: object
...
  internal_api.MonitorDTO:
    properties:
      active:
        type: boolean
      events:
        items:
          $ref: '#/definitions/internal_api.MonitorEvent'
        type: array
      history:
        items:
          $ref: '#/definitions/internal_api.HistoryPoint'
        type: array
      id:
        type: string
      interval:
        type: integer
      lastCheck:
        type: string
      latency:
        type: integer
      name:
        type: string
      status:
        type: string
      url:
        type: string
    type: object
  internal_api.MonitorEvent:
    properties:
      id:
        type: string
      message:
        type: string
      timestamp:
        type: string
      type:
        type: string
    type: object
  internal_api.OverviewResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/internal_api.GroupOverviewDTO'
        type: array
    type: object
...
  internal_api.UptimeResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/internal_api.GroupDTO'
        type: array
    type: object
info:
  contact: {}
  description: Self-hosted uptime monitoring API by Project Helena.
  title: Warden API
  version: "1.0"
paths:
...
  /monitors/{id}/latency:
    get:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Time range: 1h, 24h, 7d, 30d (default 24h)'
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_projecthelena_warden_internal_db.CheckResult'
            type: array
        "400":
          description: ID required
          schema:
            type: string
        "500":
          description: Failed to fetch latency stats
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get monitor latency history
      tags:
      - uptime
...
  /monitors/{id}/uptime:
    get:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              uptime7d:
                type: number
              uptime24h:
                type: number
              uptime30d:
                type: number
            type: object
        "400":
          description: ID required
          schema:
            type: string
        "500":
          description: Failed to calculate stats
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get monitor uptime stats
      tags:
      - uptime
...
  /overview:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.OverviewResponse'
        "500":
          description: Internal error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Dashboard overview
      tags:
      - uptime
...
  /uptime:
    get:
      parameters:
      - description: Filter by group ID
        in: query
        name: group_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.UptimeResponse'
        "500":
          description: Internal error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List monitors with history
      tags:
      - uptime
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer sk_live_..." — create keys in Settings > API Keys
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"